requirements.txt
python-dotenv==1.0.0
pandas==1.3.5
selenium==4.7.2